UBUNTU: [Packaging] Drop snapdragon config when updating from master
Running the update-from-* script will copy all config files but
remove snapdragon from being built. Running updateconfig will not
cause it to vanish but there will not be real ABI files for it which
in turn confuses final-checks because that gets its list of flavours
from the config files.

Ignore: yes
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>